refactor `get_hf_model` to shared folder
`get_hf_model` was under local_inference/util.py

moved to ./util/util.py
