+++ extensions/HuggingFace/python/src/aiconfig_extension_hugging_face/local_inference/image_2_text.py
@@ -4,7 +4,7 @@ from io import BytesIO
 from typing import TYPE_CHECKING, Any, Dict, List, Optional, Union
 
 from aiconfig.callback import CallbackEvent
-from aiconfig_extension_hugging_face.local_inference.util import get_hf_model
+from aiconfig_extension_hugging_face.util.util import get_hf_model
 from PIL import Image as img_module
 from PIL.Image import Image as ImageType
 from transformers import Pipeline, pipeline
