+++ extensions/HuggingFace/python/src/aiconfig_extension_hugging_face/local_inference/text_summarization.py
@@ -8,7 +8,7 @@ from aiconfig.default_parsers.parameterized_model_parser import (
 )
 from aiconfig.model_parser import InferenceOptions
 from aiconfig.util.params import resolve_prompt
-from aiconfig_extension_hugging_face.local_inference.util import get_hf_model
+from aiconfig_extension_hugging_face.util.util import get_hf_model
 from transformers import (
     AutoTokenizer,
     Pipeline,
